+++ rbx_binary/CHANGELOG.md
@@ -1,6 +1,9 @@
 # rbx_binary Changelog
 
 ## Unreleased
+* Changed serializer to always write a length for empty `PropertiesSerialize` values, which Roblox requires. ([#639])
+
+[#639]: https://github.com/rojo-rbx/rbx-dom/issues/639
 
 # 3.0.0 (2026-07-01)
 * Upgraded to `rbx_reflection` v3.0.0